move model-loader to config folder, update main
